requirements.txt
python-dotenv
gspread~=6.2.1
protobuf~=6.32.0
requests~=2.32.4